feat: :sparkles: add update and delet ingredient routes

+++ src/modules/drinks/useCases/updateDrinkCover/UpdateDrinkCoverService.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 import {IDrinksRepository} from '@modules/drinks/repositories/IDrinksRepository'
 import { deleteFile } from "@utils/file";
+import AppError from "@shared/errors/AppError";
 
 interface IRequest {
     drinkId: string;
@@ -15,6 +16,10 @@ class UpdateDrinkCoverService {
     ) {}
     async execute({ drinkId, coverFile }: IRequest): Promise<void> {
         const drink = await this.drinksRepository.findById(drinkId);
+        if(!drink) {
+            throw new AppError("Drink does not exit")
+        }
+
 
         if (drink.cover) {
             await deleteFile(`./tmp/drink/${drink.cover}`);
